Reload Infoblox config after deleting zone
If you delete a zone using the Infoblox backend
the zone will be deleted from Infoblox but the
zone within OpenStack will go into an ERROR state
unless an operator manually logs into Infoblox
and issues a reload of the running configuration.

This commit fixes this bug by issuing a 'restart
if needed' after deleting the zone. This does not
happen when creating zones as 'restart if needed'
can be passed through as a param into that API call.

Co-Authored-By: Jay Faulkner <jay@jvf.cc>
Change-Id: I296c5f085cce27033461be81ca58c85f095df89a
Closes-Bug: #1840253